// src/ReduxTool/Slice/Consumer/types.ts
export interface initialstateType {
    isLoading:boolean,
    error:string|null
    orderModalOpen:boolean,
    projects: ProjectTy[] | null,
    roofAnalysis: roofAnalysisType,
    ondFiles:FileUplad | null
    orderstatusmodal:boolean
};
export interface roofAnalysisType {
    isWithRoof: boolean | null;
    isSaved: boolean;
    isSaveplantSummarydata:boolean;
    selected_project:ProjectTy | null;
    esriDraw:boolean;
    obsArea:string|null;
    showConfirmBtn: boolean;
    openNoRoof: boolean;
    formDetails: roofAnalysisFormType;
    regionName:string
    countryName:string
    selectedOption:string
    subsequentOption:string
    cardIconOpen:string
}
export interface roofAnalysisFormType {
    projectSetup: TProjectTypesNew;
    quickplantAnalysis: quickplantAnalysisType;
    projectSummary: ProjectSummaryType;
    error:any;
}
export interface quickplantAnalysisType {
    environmentalImpact: EnvironmentalImpact
    costbracket:string
    currencyresdata:CurrenctResp
    electricityGenerated:any
    stringPerModule:any
    electricityGenerated2:any
    performanceRatio:any
}

export interface EnvironmentalImpact {
    carbonFootProint: string
    treesPlanted: string
}

// export interface FinancitalSaving {
//     monthlyGrossSaving: string
//     netsavings: string
//     annualSavings: string
//     paybackPeriod: string
//     pretaxIRR: string
// }

export type TProjectTypesNew = Pick<ProjectTy, 'createdId' |'plantcapacity'| 'address' |'monthlyunit'| 'consumercategoryid' | 'lat' | 'lng' | 'projectname' | 'projectid' | 'electricityrate' | 'sanctionload' | 'useablearea' | 'totalroofarea' | 'providerid' | 'projecttype'>

export type ILatLong = {
    lat:number,
    lng:number
}
export type ProjectSummaryType = {
    financial:FinRes|null,
    degradation:DegradationType|null

}

export type DegradationType = {
irr:number,
savingperiod:number
}
// ...
export type FileUplad = {  inverter: commomRes, module: commomRes }
// ...
export interface CurrenctResp{
    pkcurrencyId: string,
    basecurrency: string,
    symbol:string,
    rate: number,
    createdat: string,
    updatedat: string
}
// ...
export type FinRes = {
    montlyFinancialSaving: number,
    yearlyFinancialSaving: number,
    totalCarbonfootprint: number
    totalTrees: number
    netsavings: number
}